fix: eliminate all 5 circular dependencies
- metrics↔metrics-aggregation: extract SessionMetrics to metrics-types.ts
- telegram/types↔topic-persistence: extract TopicMeta to session-topic-types.ts
- pipeline↔state-store: extract PipelineState to pipeline-types.ts
- acp/session-service↔pause-intervention: extract AcpError to acp-errors.ts
- acp/index↔terminal-bridge: direct imports instead of barrel

Verified: TSC clean, madge clean, tests pass, -69 net lines
