+++ src/pipeline.ts
@@ -35,49 +35,9 @@ export interface BatchResult {
   errors: string[];
 }
 
-export interface PipelineStage {
-  name: string;
-  workDir?: string;
-  prompt: string;
-  dependsOn?: string[];
-  permissionMode?: string;
-  /** @deprecated Use permissionMode instead. */
-  autoApprove?: boolean;
-  stageTimeoutMs?: number;
-}
-
-export interface PipelineConfig {
-  name: string;
-  workDir: string;  // Default workDir for all stages
-  stages: PipelineStage[];
-}
-
-export type PipelineStageStatus = 'pending' | 'running' | 'completed' | 'failed';
-
-export interface PipelineState {
-  id: string;
-  name: string;
-  currentStage: 'plan' | 'execute' | 'verify' | 'fix' | 'submit' | 'done';
-  status: 'running' | 'completed' | 'failed';
-  retryCount: number;
-  maxRetries: number;
-  stageHistory: Array<{
-    stage: string;
-    enteredAt: number;
-    exitedAt?: number;
-    output?: unknown;
-  }>;
-  stages: Array<{
-    name: string;
-    status: PipelineStageStatus;
-    sessionId?: string;
-    dependsOn: string[];
-    startedAt?: number;
-    completedAt?: number;
-    error?: string;
-  }>;
-  createdAt: number;
-}
+// Re-exported from pipeline-types.ts for backward compatibility.
+export type { PipelineStage, PipelineConfig, PipelineStageStatus, PipelineState } from './pipeline-types.js';
+import type { PipelineStage, PipelineConfig, PipelineStageStatus, PipelineState } from './pipeline-types.js';
 
 export class PipelineManager {
   private static readonly PIPELINE_RETRY_MAX_ATTEMPTS = 3;
